pyproject.toml
[project]
name = "kr-forensic-finance"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Phase 1 — Beneish screen
    "opendartreader",   # DART API client
    "pykrx",            # KOSDAQ ticker list
    "pandas",
    "pyarrow",          # parquet I/O
    "marimo",           # analysis notebook format (interactive UI deferred; scripts run as plain Python)
    "plotly",           # charts in notebook
    "requests",         # WICS API calls
    "python-dotenv",

    # Cloud storage (optional — R2 upload + DuckDB remote reads)
    "s3fs>=2024.1.0",
    "boto3>=1.34.0",
    "duckdb>=1.0.0",

    # Phase 2 additions (uncomment when starting Phase 2):
    # "dart-fss",       # DS005 CB/BW structured endpoint helpers
    # "playwright",     # SEIBRO scraping (WebSquare JS rendering)
    # "scipy",          # statistical tests for timing anomaly analysis

    # Phase 3 additions (uncomment when starting Phase 3):
    # "networkx",       # officer network graph
    # "pyvis",          # interactive HTML graph rendering
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["02_Pipeline", "03_Analysis"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
